Print freeFunction symbolically
.. when NativeAllocationRegistry freeFunction invocation times out.

We use a thin wrapper around dladdr() to do so.

Note that in practice this only prints the dynamic library name, since
the free function is typically declared static, and thus the symbol is
not visible to dladdr. Thus it may still be necessary to disassemble the
library to debug.

Flag: EXEMPT bugfix
Bug: 417713598
Test: Temporarily add freeFunction log spam, and check output.
Change-Id: I1d767a4789f8ec58320160e80a117865c0eae079