Throw exceptions for DCL that is not read only
The code ensures that native libraries are loaded only if read-only. If this is not the case and an exception is not thrown, it logs to WW in order to be able to determine if this happens in the field.

Bug: 408070506
Test: Run test_framework_dcl
Flag: com.android.art.rw.flags.read_only_dynamic_code_load_ww_log
Flag: com.android.art.rw.flags.read_only_dynamic_code_load_throw_exception

Change-Id: I6c154017ab4db89f29f2761d141a3b509d5987eb