Add previously released API to art-module-public-api-stubs-nullability-validation
This is needed because Metalava is going to be strict about requiring a
previously released API to ensure that flagged APIs work correctly.
While `art-module-public-api-stubs-nullability-validation` does not use
flagged APIs or provide compatibility guarantees on its API but it is
also using `droidstubs` in a very different way to every other use in
the build and it is not worth special casing it.

Bug: 425309457
Test: TH
Change-Id: I0188454f15cfe4f026808cb6c9c6646c82bd8315