Add previously released API to art.module.toolchain.api
This is needed because Metalava is going to be strict about requiring a
previously released API to ensure that flagged APIs work correctly.
While `art.module.toolchain.api` does not use flagged APIs or provide
compatibility guarantees on its API it does need a previously released
API and it is not worth adding in code to the build to special case it.

Bug: 425309457
Test: TH
Change-Id: I388fab25d0a5482e2cb0094c64e227b7fe41b4b5