Maintain separate niceness and priority
We decided to keep independent priority and niceness values in Thread,
with niceness closer to the actual Linux niceness value, but not
inherited by Java threads, and not affecting Thread.getPriority().
This makes the necessary ART changes.

This is largely forced by the realization that
Process.setThreadPriority() values were not inherited, and clients
probably rely on that.

This is the libcore part of that change.

Flag: EXEMPT bugfix
Bug: 417492941
Test: Treehugger
Change-Id: Ib58f057bad2a2640650be24e675b1abf4291c1ad