Add a core-lambda-stubs-src that is always built from source for use in SDK builds
core-lambda-stubs can be replaced by the prebuilt from the ART module SDK
depending on the product config.  This can result in a mismatch between
the Linux and Mac SDKs that apparently causes issues.  Create a copy of
core-lambda-stubs that is always built from source for use in the SDK
build so it gets consistent results.

Bug: 237299698
Bug: 440324010
Test: builds
Flag: EXEMPT REFACTOR
Change-Id: I67028b99d09babc013cffac483ca9c567a979bf9