Remove launched flag com.android.libcore.schedule_at_fixed_rate_new_behavior
The new behavior is now applied if SDK_INT >= 31, via ART Mainline updates.
This means that we can remove related details in the Javadoc, because we're essentially masking the freezer-related issue when this new behavior is enabled.
There is no need for the ChangeId anymore. There is no toggle on app targetSdkVersion, and there is no expectation to remotely disable this behavior.

Flag: EXEMPT FLAG_REMOVAL
Fixes: 452739244
Change-Id: I0411a76e15f2435a4a2da19e8efb54b6df8374da