Change logger level/parent only when necessary
Only change logger level when the requested level is different
than the current one.

This is also true for Logger.setParent().

e.g. when a new process is forked, logger level will be reset
to a default, and this is usually the log level before a fork.
This hange avoids a COW page fault/allocation due to a write
that doesn't change anything for each fork.

Bug: 457449920
Flag: EXEMPT - bugfix

Change-Id: I02046fe5f2abba4e837b001fe2f850db20fba215