Add HashMap to avoid unexpected interface error
In the main `ojluni` sources the `HashMap` and `LinkedHashMap` form a
dependency cycle between their nested classes, i.e.
* `HashMap.TreeNode` extends `LinkedHashMap.Entry`
* `LinkedHashMap.Entry` extends `HashMap.Node`
* `HashMap.Node` implements `Map.Entry`

The `ojluni/annotations/mmodule` directory contains a
`LinkedHashMap.annotated.java` class that does not have a nested
`Entry` class, so when resolving the super class of `HashMap.TreeNode`
it resolves to `Map.Entry` which is an interface which is not a valid
super class.

This change copies `HashMap.annotation.java` from the
`ojluni/annotations/flagged_api` directory to break the cycle and avoid
trying to use `Map.Entry` as a super class. This is preparation for
future Metalava changes that improve the error checking.

Bug: 458362596
Test: ./gradlew
Change-Id: If9802fce1d7de0a86c61b74082f216ec1bf35565