Avoid non-block tag uses of `@hide`
Previously, Metalava ignored non-block uses of `@hide` when they were
used in an comment that also had a block use of `@hide`. That was
because it had no effect. However, in order to fix an issue where
system APIs were not being included in system API documentation (see
bug for details) it now does change and report issues with them.

Rather than try and treat this as a special case or baseline it this
change replaces the `@` in `@hide` with `&#64;`. It also tweaks the
wording to avoid reformatting the containing paragraph.

Bug: 462333324
Test: ./gradlew
Change-Id: Iaee2c814846b17ed63e19964f25aa66b88a0464d