Make dexTypeIndex non-volatile
There are no Java uses, and native accesses are not `volatile`, so the
`volatile` here, and the associated comment, only misleads. I found no
indication that problem it attempts to solve is real.

The `volatile` here does not appear to have had any effect on generated
code.

Flag: EXEMPT BUGFIX
Bug: 472336923
Test: Treehugger
Change-Id: I386fd67b4f421067e4cebec0bde6a4b1356fe286