Rename Class.clinitThreadId
It was marked as

 @UnsupportedAppUsage(maxTargetSdk = 30, trackingBug = 170729553)

and was disallowed in https://developer.android.com/about/versions/12/non-sdk-12
as having little or no usage.

Since we're changing how the field is maintained, it seems safer to
rename it to be consistrent with the new ART name. The semantics were
and are obscure enough to make correct use in app code very unlikely,
either before or after the change. (Previously it was cleared when
written to an image. Now it also changes meaning once the class is
initialized.)

Flag: EXEMPT BUGFIX
Bug: 454767701
Bug: 458973673
Test: Treehugger
Change-Id: I7089e89d39a9ba52fa4f5e3a61a70983a0ba9494